requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
langchain==0.1.7
langchain-core==0.1.15
langchain-google-genai==0.0.9
langchain_text_splitters==0.3.8
google-generativeai==0.3.2
beautifulsoup4==4.13.4
chromadb==0.4.22
mysql-connector-python==8.3.0

# 테스트 관련 패키지
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2 